+++ HISTORY.rst
@@ -2,6 +2,8 @@
 ------------------
 
 - Allow marking a note as containing HTML. (#17)
+- Add languages attribute to :class:`skosprovider.skos.ConceptScheme` to make it
+  possible to track what languages are being used in a thesaurus. (#19)
 
 0.5.3 (2015-06-24)
 ------------------